+++ app/src/main/java/com/elogictics/taskmaster/model/request/AddMaterialRequest.kt
@@ -3,7 +3,7 @@ package com.elogictics.taskmaster.model.request
 data class AddMaterialRequest(
     private val mateId: Int,
     private val jobId: Int,
-    private val weight: Long?,
-    private val weightToCus: Long?,
-    private val price: Long?,
+    private val weight: Double?,
+    private val weightToCus: Double?,
+    private val price: Double?,
 )
